refactor buy and sell full set modals to declarative modal pattern
