+++ components/markets/FullSetButtons.tsx
@@ -2,36 +2,14 @@ import { Dialog } from "@headlessui/react";
 import Modal from "components/ui/Modal";
 import { useMarket } from "lib/hooks/queries/useMarket";
 import { useMarketIsTradingEnabled } from "lib/hooks/queries/useMarketIsTradingEnabled";
-import { useModalStore } from "lib/stores/ModalStore";
 import { observer } from "mobx-react";
 import dynamic from "next/dynamic";
 import { useState } from "react";
 
-import BuyFullSetModal from "./BuyFullSetModal";
-import SellFullSetModal from "./SellFullSetModal";
+import BuyFullSetForm from "./BuyFullSetForm";
+import SellFullSetForm from "./SellFullSetForm";
 
 const FullSetButtons = observer(({ marketId }: { marketId: number }) => {
-  const modalStore = useModalStore();
-  const modalOptions = {
-    styles: { width: "304px" },
-  };
-
-  const handleBuyFullSetClick = () => {
-    modalStore.openModal(
-      <BuyFullSetModal marketId={marketId} />,
-      <>Buy Full Set</>,
-      modalOptions,
-    );
-  };
-
-  const handleSellFullSetClick = () => {
-    modalStore.openModal(
-      <SellFullSetModal marketId={marketId} />,
-      <>Sell Full Set</>,
-      modalOptions,
-    );
-  };
-
   const { data: market } = useMarket({ marketId });
   const enabled = useMarketIsTradingEnabled(market);
 
@@ -61,13 +39,19 @@ const FullSetButtons = observer(({ marketId }: { marketId: number }) => {
 
       <Modal open={buyModalIsOpen} onClose={() => setBuyModalIsOpen(false)}>
         <Dialog.Panel className="w-full max-w-[462px] rounded-[10px] bg-white">
-          <BuyFullSetModal marketId={marketId} />,
+          <BuyFullSetForm
+            marketId={marketId}
+            onSuccess={() => setBuyModalIsOpen(false)}
+          />
         </Dialog.Panel>
       </Modal>
 
       <Modal open={sellModalIsOpen} onClose={() => setSellModalIsOpen(false)}>
         <Dialog.Panel className="w-full max-w-[462px] rounded-[10px] bg-white">
-          <SellFullSetModal marketId={marketId} />,
+          <SellFullSetForm
+            marketId={marketId}
+            onSuccess={() => setSellModalIsOpen(false)}
+          />
         </Dialog.Panel>
       </Modal>
     </div>
